requirements.txt
eventlet==0.18.4
Flask==0.10.1
greenlet==0.4.9
itsdangerous==0.24
Jinja2==2.8
MarkupSafe==0.23
python-engineio==0.8.8
python-socketio==1.0
six==1.10.0
Werkzeug==0.11.4
wheel==0.26.0
